add login and sign up

+++ controllers/feed.js
@@ -3,6 +3,8 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
+const user = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -46,21 +48,27 @@ exports.createPost = (req, res, next) => {
   }
 
   const imageUrl = req.file.path.replace("\\", "/");
-
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'Deepak'
-    },
+    creator: req.userId,
   })
   post.save()
+    .then(() => {
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
     .then(result => {
-      console.log(result)
       return res.status(201).json({
         message: 'post created successfully!',
-        post: result
+        post: post,
+        creater: { _id: creator._id, name: creator.name }
       })
     })
     .catch(err => {
@@ -121,6 +129,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!')
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl)
       }
@@ -153,8 +167,20 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!')
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId)
+    })
+    .then(() => {
+      return user.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId)
+      return user.save();
     })
     .then(result => {
       res.status(200).json({ message: 'Post deleted!' })
